requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-allauth==65.10.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
django-environ==0.10.0
Pillow==10.3.0
